style: Update button styles and add global CSS (#11)
Button styles in LogView.tsx were updated to ensure a more visually appealing interface. A stack was also adjusted to have a minimum width. Moreover, a global CSS file, `index.css`, was added along with typography settings in `main.tsx` which included specifications on font family, line height and color schemes to help improve the overall appearance of the application.